flask-app/requirements.txt
elasticsearch==7.17.0
Flask>=2.0.0
Flask_MySQLdb==1.0.1
mysqlclient==2.1.0
python_bcrypt==0.3.2
requests>=2.27.1
